Move installation functionalities from `Command` to `CommandBuilder`
