+++ src/statue/command_builder.py
@@ -1,6 +1,12 @@
 """Build commands from configuration."""
+import importlib
+import os
+import subprocess  # nosec
+import sys
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional
+
+import pkg_resources
 
 from statue.command import Command
 from statue.constants import (
@@ -14,6 +20,7 @@ from statue.constants import (
 )
 from statue.context import Context
 from statue.exceptions import InconsistentConfiguration, InvalidCommand
+from statue.verbosity import DEFAULT_VERBOSITY, is_silent
 
 
 @dataclass
@@ -145,6 +152,38 @@ class CommandBuilder:
     )
 
     @property
+    def install_name(self) -> str:
+        """
+        Name to state while installing with pip.
+
+        When installing a specific version with pip, one should add "==" with the
+        specific version afterwards.
+
+        If no version is specified, same as name.
+
+        :return: name and version
+        :rtype: str
+        """
+        if self.version is None:
+            return self.name
+        return f"{self.name}=={self.version}"
+
+    @property
+    def installed_version(self) -> Optional[str]:
+        """
+        Version of the installed package.
+
+        Might not be the same as the version attribute.
+
+        :return: version of installed package
+        :rtype: str or None
+        """
+        package = self._get_package()
+        if package is None:
+            return None
+        return package.version
+
+    @property
     def specified_contexts(self) -> List[str]:
         """Contexts names list with arguments specifications."""
         return list(self.contexts_specifications.keys())
@@ -157,6 +196,110 @@ class CommandBuilder:
             *self.allowed_contexts,
             *self.specified_contexts,
         ]
+
+    def installed(self) -> bool:
+        """
+        Is this command installed.
+
+        :return: Either the command is installed or not
+        :rtype: bool
+        """
+        return self.installed_version is not None
+
+    def installed_correctly(self) -> bool:
+        """
+        Checks that command is installed and its version matches.
+
+        :return: whether the command is installed correctly
+        :rtype: bool
+        """
+        return self.installed() and self.installed_version_match()
+
+    def installed_version_match(self) -> bool:
+        """
+        Is the installed version match the specified version.
+
+        :return: is the installed version matches the desired version
+        :rtype: bool
+        """
+        if self.version is None:
+            return True
+        return self.installed_version == self.version
+
+    def install(self, verbosity: str = DEFAULT_VERBOSITY) -> None:
+        """
+        Install command using pip.
+
+        :param verbosity: Verbosity level.
+        :type verbosity: str
+        """
+        if self.installed():
+            return
+        if not is_silent(verbosity):
+            print(f"Installing {self.install_name}")
+        subprocess.run(  # nosec
+            [sys.executable, "-m", "pip", "install", self.install_name],
+            env=os.environ,
+            check=False,
+            capture_output=is_silent(verbosity),
+        )
+
+    def update(self, verbosity: str = DEFAULT_VERBOSITY) -> None:
+        """
+        Update command using pip.
+
+        :param verbosity: Verbosity level.
+        :type verbosity: str
+        """
+        if not is_silent(verbosity):
+            print(f"Updating {self.name}")
+        subprocess.run(  # nosec
+            [sys.executable, "-m", "pip", "install", "-U", self.name],
+            env=os.environ,
+            check=False,
+            capture_output=is_silent(verbosity),
+        )
+
+    def uninstall(self, verbosity: str = DEFAULT_VERBOSITY) -> None:
+        """
+        Uninstall command using pip.
+
+        :param verbosity: Verbosity level.
+        :type verbosity: str
+        """
+        if not self.installed():
+            return
+        if not is_silent(verbosity):
+            print(f"Uninstalling {self.name} (version {self.installed_version})")
+        subprocess.run(  # nosec
+            [sys.executable, "-m", "pip", "uninstall", "-y", self.name],
+            env=os.environ,
+            check=False,
+            capture_output=is_silent(verbosity),
+        )
+
+    def update_to_version(self, verbosity=DEFAULT_VERBOSITY) -> None:
+        """
+        Update command to the specified version using pip.
+
+        If the installed version is the same as version, do nothing.
+
+        :param verbosity: Verbosity level.
+        :type verbosity: str
+        """
+        if not self.installed():
+            self.install(verbosity=verbosity)
+            return
+        if self.version is None:
+            # If no version is specified, we update package to its latest version
+            self.update(verbosity=verbosity)
+            return
+        if self.installed_version_match():
+            return
+        # If a version is specified, we must first uninstall it
+        # before installing the specified version.
+        self.uninstall(verbosity=verbosity)
+        self.install(verbosity=verbosity)
 
     def validate_contexts(self, *contexts: Context):
         """
@@ -306,3 +449,17 @@ class CommandBuilder:
         )
         command_builder.update_from_config(builder_setups)
         return command_builder
+
+    def _get_package(self):  # pragma: no cover
+        """
+        Get package of the desired command.
+
+        If package is not installed, returns None.
+
+        :return: self package
+        """
+        importlib.reload(pkg_resources)
+        for package in list(pkg_resources.working_set):
+            if package.key == self.name:
+                return package
+        return None
